Revert "Add handling of field config in fieldDropdown (#51750)"
This reverts commit 4f373c474ebe31623fe49034938bb29bdac75a73.
