+++ apps/i18n/common/en_us.json
@@ -2356,6 +2356,7 @@
   "unassignSectionConfirm": "Your students in {sectionName} will no longer be taken to {courseName} when they sign in.",
   "unassignedStudents": "Unassigned Students",
   "uncompletedLessons": "Lesson not completed",
+  "uninitialized": "uninitialized",
   "unsubmit": "Unsubmit",
   "unsubmitAssessment": "Unsubmit your assessment",
   "unsubmitYourProject": "Unsubmit your project",
