+++ apps/src/blockly/constants.js
@@ -46,6 +46,3 @@ export const CLAMPED_NUMBER_REGEX =
 
 // Used for custom field type FieldButton for 'play sound' block
 export const DEFAULT_SOUND = 'sound://category_digital/ping.mp3';
-
-export const NO_OPTIONS_MESSAGE = 'uninitialized';
-export const EMPTY_OPTION = '???';
